Update README.md for AppImage and possible Windows/macOS errors
For this upcoming version, I've overhauled my release process, upgraded
the bundled LÖVE to 11.4 (64-bit versions in all cases) and will
include AppImages for Linux (also so that I can release on itch.io!)

The README needed a bit on an update, so for Linux I've added info
about the AppImage and that you should not install LÖVE via Snap (which
used to not really be a thing but now it is). For Windows, I've added
some info about the 64-bit version, and for Windows and macOS I've
added info about the typical "Windows protected your PC" and "developer
cannot be verified" messages that people may get.

[pre45]