Fix decoding XML entities above &#x7f;
XML entity decoding used string.char(), which doesn't support Unicode.
This is *normally* not a problem - XML entities are normally only used
to encode control characters, which are equivalent between ASCII and
UTF-8, and I'm pretty sure both Ved and VVVVVV never create higher XML
entities. But they're perfectly valid, so someone might create one, and
that should not cause �s or even crashes when you use &#x100; or above.
(I'm also thinking of a change that would need &#xA4; to work but
that's a long story to explain.)

[pre02]