Fix "Unrecognized command" showing up for text box lines and comments
5046ea7105a5f825ce13c0fd7e1963ed0f7f3763 added an explicit warning text
if your cursor is on a line with an invalid command or a command in the
wrong (simplified/internal) language. However, this didn't make a
distinction between command lines and non-command lines - so you'd
often see the warning text for lines that are part of text boxes, as
well as lines starting with # or // (which are seen as comments). This
commit stops that from happening.

[pre03]