Fix alt+left/right also affecting cursor position in script editor
Alt+left/right navigates through scripts or creates new scripts, based
on lines referencing scripts. Left/right also goes to the previous/next
character in the input system, which does not require modifier keys to
not be held. So if you press alt+right, you may also go to the next
line in the current script, which is most noticeable if you press
alt+right to create a new script but not go to that new script yet.
This was reported by Misa on Discord.

This highlights another issue: on macOS, alt+left/right is already used
to navigate by whole words (which is ctrl+left/right on Windows and
Linux.) So the alt+left/right shortcut for navigating/creating scripts
needs to be changed to cmd on macOS.

[pre02]