Fix shift+left/right sometimes canceling selection
As reported by Ally on Discord, if you press left/right at the
start/end of a line while holding shift to change a selection, the
selection would, in some circumstances, completely disappear. This was
because moving between different lines happened in two steps: an
input.movey to move to the other line, and then an input.setpos to set
the X position to the start/end of that line. The first step would
cancel the selection based on the cursor position between the two
steps - the new line, but with the old X position. This is now fixed by
using input.setpos for both the X and Y.

[pre03]