Add partially working font editor
This is more of a prototype right now, since it doesn't have full
functionality yet - but much of the interface is now there.

First commit in a while, for multiple reasons:
- For a lot of the time I was working on development of VVVVVV itself,
  especially its new font system with support for CJK and button
  glyphs. And I also went skiing for a week at some point
- The font editor changes in this commit have been unstaged for months,
  so it has also been kinda holding me back from making quick commits -
  there was a bit of a barrier that I needed to clean this up first
- In the background, I have also been working on a brand new XML
  library for Ved, which will solve multiple issues - the existing
  pattern matching system really needs to be replaced, with all the new
  XML that Ved will need to handle (new 2.4 tags in level files,
  fontmeta, etc). This project is also looking good!

Feels good to finally commit this and get back to a clean working
directory! It's in a separate branch, so I can release smaller versions
before 1.11.0, which is probably necessary.