Swap "Assets" and "Level notepad" buttons
It's kind of a weird sight having that button be a really short label!

It makes sense to give the Assets button some more prominence though.
Some people didn't even know that button was there (judging from a
conversation on Discord), even though custom assets are pretty
important for many people nowadays (that's why the level-specific
assets screen was added).

Level notes seem more rarely used. The only reason Assets was on the
second page was because it was added later than the level notes.

[pre04]