Don't interpret Ctrl+Shift+N as "new level" in main editor
We had a conflict between the following shortcuts:
- Holding Ctrl+Shift to display all tiles
- Holding N to show tile numbers (including in the tiles picker)
- Pressing Ctrl+N to start a new level

Ctrl+N would be triggered even if Shift was held, thus making holding
down Ctrl+Shift and N to display tile numbers in the tiles picker
impossible. This is now fixed.

This would also be fixed by the customizable keybindings system I'm
planning, but for now hardcoding it will do.

[pre08]