Extend script editor coordinates preview with clicking to set coords
Clicking in the small canvas changes the coordinates in the script line
itself (and you can also click-and-drag to slide the coordinates to the
value you want)!

This probably needs improvements later. For example, this could be
optimized to put gotoposition/createcrewman aligned so the crewmate
will stand on the floor instead of in mid-air aligned to a 4x4 grid,
and maybe the canvas is a bit too small to comfortably set positions,
but it may still be really useful already.

[pre11]