Turn JIT compilation off, temporarily
For some bizarre reason, on the same version that you'd think was tried
and tested for last year's contest levels, multiple people are now
reporting occasional unexpected freezes with seemingly no cause. It's
rare enough that it's basically impossible to reproduce, and common
enough that it's an absolute PITA - every freeze is one too many.

The most likely cause - I think - is some kind of bug in the LuaJIT
version that comes with LÖVE 11.4. Looking at the LÖVE Discord, there
*are* certain instabilities (apparently especially `pairs` can
sometimes freeze, misbehave or crash?) and those bugs have since been
fixed in a newer version of LuaJIT, but we're not getting those fixes
until LÖVE 11.5.

Seemingly, jit.off() helps make things more stable, so let's try that!
If there's no reports of freezes anymore, it doesn't really say
everything because it was that rare for it to happen, but let's hope
there won't be any reports anymore. I'll update to LÖVE 11.5 as soon as
it comes out.

This may make Ved a bit slower, but it shouldn't be too dramatic.

[pre12]