Handle right-to-left and/or bottom-to-top bounds placing better
This is still not perfect. The platform and enemy bounds could already
be placed into any direction, but there was no protection against
getting "out of order" corners (like making the bounds 1 tile
wide/high). I don't know how VVVVVV handles these reversed bounds, but
when placing bounds out of order in VVVVVV, at least in 2.4, it always
fixes the first coordinate to be top left and the second coordinate to
be bottom right. So Ved now does that too: x1,y1 will always become top
left, and x2,y2 will always become bottom right.

Ideally, while you place the second corner, the bounds follow your
cursor like they do when placing script boxes, but since they currently
rely on temporarily setting the second coordinate to 320,240, that's a
bit more annoying to do. It should probably be a part of the tools
overhaul to do at some point: it kinda makes sense to make changing the
boundaries a (hidden?) subtool of the platforms/enemies tools, so it
can *hopefully* rely on the same code as script boxes will. And the
current bounds-changing code (relying on the variable editingbounds)
already ties into the platforms/enemies tools, so...

[pre14]