Check jit ~= nil before calling jit.off()
This is probably not necessary - I coincidentally noticed LÖVE 0.8.0
doesn't have the jit module and this line makes Ved('s error screen)
completely and silently fail to open with no message whatsoever. And
checking for jit being nil fixes it. So just in case the jit module is
missing in some real-life scenarios, let's just skip turning jit off in
that case.

[pre15]