Fix symlinked plugin directories not loading
Due to a bug in LÖVE that goes down to PhysFS[1], a symlinked directory
isn't recognized as a directory, so symlinked plugin directories won't
be loaded.

To fix this, just load it anyway if it's a symlink. If the symlink turns
out to be a file, that's harmless because it will just be loaded as an
unknown plugin that does nothing.

[1]: https://github.com/love2d/love/issues/1938