Fix PageUp and PageDown keys in script editor
This was broken in 2da88b62818d1f8d22788f08ef513c35441809ab (1.10.2),
because it assumed if the Y position went out of bounds, it meant you
had pressed Up on the first line or Down on the last line, and set the
cursor position to leftmost/rightmost accordingly, without saving a new
Y position. Now these shortcuts work as expected again.

[pre02]