Fix line flashing in script editor if you click too low on a line
This took a while to debug, turns out there were two places in the code
that were setting cursor position on click: input.drawcas (because
that's how the new input system works) as well as leftover code from
before the new input system. It was also calling input.setpos on click
(I did migrate that over but I forgot it had become completely
redundant) and it sometimes disagreed on the exact coordinate.

I love how the solution ended up being:
    git rm uis/scripteditor/mousepressed.lua

[pre03]