Fix anythingbutnil0 now possibly returning nil
A few commits ago, a change was made so that if the input wasn't nil,
instead of returning the input as-is, it was passed through tonumber
to force it to be a number. Unfortunately, tonumber can return nil of
course, and I didn't think of that, so this was causing some new
crashes. This is fixed now.

[pre04]