Rewrite autotiling completely to match VVVVVV 2.4
Autotiling was a mess of functions and if chains and switch statements.
This commit makes autotiling better by assigning each direction to one
bit in a byte, giving each different combination its own value.
This value is then fed into a lookup table to give fine control on
which tiles get placed where.

The lab tileset can now use the single tiles which were before unused
in the autotiler, and the warp zone's background tool now places the
fill used in the main game.

This matches 2.4.