Fix high CPU usage when holding N to print tile numbers
Holding N caused high CPU usage, because printing every tile number
used up to 24,000 prints _every_ frame, due to the color fading style
and the shadows needing five prints for each digit. The text system
was not designed for that - specifically, the point of SpriteBatches
was subverted since every print needs its own batch, so each batch had
only one sprite in it.

The solution was to not use standard prints for the tile numbers, but
to make special separate SpriteBatches just for those. This system
can be found in tilenumberbatch.lua. There's also a cache of what
numbers are displayed so the SpriteBatches don't need to be re-filled
every frame.

[pre08]