Keep around 2.4 <TextboxColours> and <SpecialRoomnames>
Just like <font>, levels may have these tags in 2.4, so until we can
fully support them, we should keep them around.

In this case, a complicating factor is that Ally's Future plugin also
added code that carries these tags. So in order to avoid breaking
levels by having an old version of Future installed in this Ved, I'll
break a code replacement that Future used on purpose, and for now I'll
version that bit of code so I won't have to rename a variable again
(and I'm submitting a PR to update Future for this as well)

[pre14]