Fix crash screen crash if plugin has early error
If a plugin had a syntax error or something, and thus caused a crash
at a really early point, the scale setting (s.pscale) would not be
initialized yet, and since that was used in one place in the crash
screen without checking, the crash screen would crash and result in
the fallback crash screen to show up. This is now fixed - now the
normal crash screen will show.

[pre15]