Don't show zip option in certain save dialogs
If you're quitting Ved and saving unsaved changes, and use the option
to also save the level as a zip, the zip dialog doesn't appear because
we quit. When starting a new level it just makes a dialog instantly
appear and disappear and then it says it can't create the zip file.
So it's better not to offer the option to save a zip file in those
cases.

[pre16]