Keep last input when returning to load screen
When searching for a level name in the levels list, then going into
the main editor (for example, by selecting a filtered result), and then
going back to the levels list, the list would still be filtered, but
the input field would be empty. To clear the leftover filter, you would
need to type a letter and then remove it again (or refresh the list).

Now, when you return to the levels list, the old input is filled back
in, so you can clear it more easily by pressing Ctrl+Backspace for
example. Or you can keep working with it, since it may actually be
really convenient to be able to easily load different versions of the
same level in succession.

[pre16]