Remove F9 hotkeys being hidden if editing roomname
It was a nice touch, but it makes the code more nightmarish.
Many conditions looked like:
  (not editingroomname and {L.PLATFORMBOUNDS, "t"} or L.PLATFORMBOUNDS)

There were also forgotten ones which would be shown anyway (probably
due to the fact that this was so hard to have an overview of), and for
some reason, where editing the roomname was thought of, editing
roomtext was not, even though the shortcuts were disabled in the same
way as far as I can see?

Anyway, maybe this will come back later if the hotkey displaying
functionality can dynamically recognize when text input is currently
active and which shortcuts to show and which ones to hide, but for now
it's probably not worth it.

[pre17]