Remove all remaining _= conditional statements
These looked like:
	_= condition and statement()

Which could instead be something like:
	if condition then
		statement()
	end

I'm pretty sure the rationale behind these was when I needed lots of
lines in a row (for example when displaying a lot of UI elements below
each other or something) of which many, scattered throughout, needed to
be conditional to something, and others in the middle of them didn't.
But it's ugly, and I just already removed most of them in the previous
few commits, so I just removed all the remaining ones as well.

[pre17]