Fix off-by-one in gravity line solidity checks
When placing a horizontal gravity line, we were potentially checking
tiles 0-40 for solidity, not 0-39. Before a recent change that made
roomdata_get(...) strict on requesting a tile outside the room
boundaries, that went unnoticed because it would just get the first
tile on the next line (and on the last line it would just be a nil
tile, haven't checked but it didn't result in a crash either).

This all meant that if you let a gravity line go offscreen on the
right, it would only stick out (like it's supposed to) if the first
tile on the line below would let gravity lines through; otherwise it
would stop at the room border. This might not result in any practical
differences...? (for example, missing the line in certain pixel-
specific scenarios) but it's good that this is now also fixed anyway.

I also did some minor cleanups to the function.

[pre20]