Fix `partss_parsed` nil crash
This makes just_text(...) consistent with syntax_hl(...).

How I explained it on Discord (edited a bit):

The problem has to do with `textsprite` going above the screen border.
It doesn't need to color the line anymore, but it still wants to know
if a long textbox started somewhere offscreen that should continue
onscreen.

Since the line starts with `text`, it tries to take more effort to see
if it's actually a "valid" `text` command, and since it's not a `text`
command, it tries to see if it's instead `setactivitytext`, but that
part of the code references a variable that does not exist.

[pre04]