Remove usefontpng and uselevelfontpng options
There are two font-related options in Ved: the option to use the font
from the graphics folder, and the option whether to use the font from
level-specific assets or not.

The first option made more sense back when the only way to customize
assets for a level was to put them into your global VVVVVV folder.
VVVVVV also only supported a limited ASCII set which meant Ved's
built-in font had support for characters that couldn't be supported in
VVVVVV's font (which affected Ved's interface if you used a different
language in Ved). Since 2.3, there isn't much reason to customize your
VVVVVV font globally anymore other than just wanting to style your
VVVVVV differently for some reason.

The second option is partially the same story, but the other way
around: it's not really necessary anymore to turn it off when Ved uses
a different font for the interface and level content, which is what I'm
working on right now. Actually I can think of one reason to turn it
off: there's one level with a blank font.png on purpose, which
currently makes Ved's interface unreadable. So that problem will
disappear too.

Now that we need to overhaul the font system for 2.4, especially now
that fonts can have different sizes than 8x8, removing both options
makes things a little bit simpler to implement. Maybe one or both will
come back later, but for now, let's see that everything works properly.

[pre05]