Allow level font selection
There's now a dropdown in the advanced level options which allows you
to select the font. (Maybe it'd be better in the regular level options,
but for now it'll do).

This required adding support for per-option fonts in dropdowns and
right click menus. I'll also leave you to imagine the amount of
confusion and debugging which preceded the change to table.copy, when
cVedFont objects only had part of their attributes and none of their
methods...

[pre06]