Improve text createcrewman red channel consistency
What that mouthful means is that, since 2.4, the r can be set to more
arbitrary values using createcrewman by passing in a number for the
color. And to make that be consistent in Ved, I had to do some
refactoring, which had the side effect of further improving the code.

The alttextboxcolors table is now removed, since it's now based on
dynamically changing the red channel from a textboxcolors, based on
what the third argument to a createcrewman command is (including when
the argument is _not_ passed and thus an argument from a previous
comment is reused).

Custom <TextboxColours> are still not displayed, but once we have the
data it would be a pretty simple addition to the newly added
set_textbox_color(...), which also helped reduce code duplication a
little bit (the same textbox coloring code was in two places).

Actually, <TextboxColours> is a simple 2D data structure, so might as
well parse and store it _before_ finally implementing that new XML
parser...

[pre08]