Allow element containers to have iterators
In addition to rigid tables, lists of elements (like ListContainer,
but also the root elements list of a stat) can now also be specified as
an iterator function. This can result in a bit more flexibility, even
though it has some big caveats related to not creating new objects
every frame... But it may still be useful at some point.

(Actually I _was_ going to use it to dynamically generate a list of
elements, before realizing that wasn't a good idea. But I can also
think of usecases where it may be useful, like the iterator function
automatically updating a cached list of elements depending on whether
they're stale or not, before iterating over it - so I'll keep it.)

[pre09]