Fix first frame flashing in elements system
In the elements system, the placement of elements may be dependent on
the dimensions of other elements which aren't known yet before the UI
is actually drawn, which means you may see some kind of flash when
opening a state for the first time. This was already noticeable in the
"VVVVVV setup" screen, for example. It now got especially noticeable in
the new textbox colors screen, since the layout can be refreshed while
you're working with it, and all the color names will twitch for a
single frame as they have to be vertically centered before it is known
how high the text will be.

The solution is pretty simple: uis are now given a `drawn` attribute,
the lack of which signals that the UI has not yet been drawn. It will
therefore need a number of frame draws to be made invisibly before
the elements have settled in place. So, we turn off all color
components so that nothing affects the screen (I could've also done a
canvas, something offscreen, a 0-pixel scissor...) and then draw the
UI a number of times in a row. The result is completely unnoticeable -
after all, the UI is just "drawn on top of itself" a few times before
actually drawing it visibly within the same frame.

What number of times? I think given a large enough balancing act of
elements that keep going further and further down, this number could be
indefinite? But throwing away a single frame already seems to fix all
the problems I could find, so 2 should be even better to be sure. And
2 is also still pretty harmless. If there's still a problem later, we
can always increase it a bit more.

Of course, it's always possible to set this `drawn` attribute to false
to force the next few frames to be thrown away, and this is done in the
textbox colors screen when the list of colors is changed.

[pre10]