Migrate scripting syntax color options to elements
With adding ColorButton and ColorPicker elements for the new textbox
colors screen, which duplicated the code of the syntax highlighting
color settings screen, the plan was to migrate the syntax highlighting
settings to the elements system as well, so that's now done! Once again
a net cleanup - it's satisfying to remove functions that only existed
to run the same code when clicking a button or pressing a hotkey.

I did have to fix [Horizontal]ListContainer elements to give a pw/ph
equal to the actual space it takes in if there's no bottom/right
elements, instead of always both the maximum width and height given to
it. So that's also a good result.

[pre11]