Remove CRLF conversion for clipboard empty check
Replacing \r\n by \n will never change whether or not the clipboard is
an empty string, so it's not really necessary to do so. I think it's
done anyway for symmetry with the other love.system.getClipboardText()
where the text is actually pasted, to ensure if any other conversions
are added (stripping whitespace or something), the presence of the
"Paste" option doesn't mismatch whether text being can be pasted. So I
added a comment to be sure.

[pre12]