Add support for bidi text rendering
This was not too hard, since I already had the same thing working in
VVVVVV, so I could just reuse most of the same code with only small
modifications. This change adds a C library to Ved, which is not the
first time, but this is the first time that I'm going to ship DLLs
for Windows as well.

Speaking of which, I haven't included any of the binaries in this
commit yet, because I'm still probably going to make changes to handle
TEXT INPUT as well as possible. This is kind of the final boss of bidi
support, especially with selection and stuff like that. But I think
it's realistic? Let's just see...

[pre13]