Add support for bidi text input
Okay, I've always thought *displaying* bidirectional text would already
be incredibly difficult to support (I mean, fair enough, just use an
external library for the most difficult part), but text input (beyond
just sticking a "_" at the end of the string and not allowing the user
to move the cursor throughout the string) would be next level.
Discontinuous selections especially?

But I decided to give it a go anyway, and it turned out easier than I
expected. Since we already have bidi rendering in place, we already
have access to all the information we need: which character from the
original in-memory string changed to which in the reordered string,
which glyphs are in LTR runs and which in RTL runs, and how wide is
each displayed glyph. If you have that info, and you sorta know how
bidi works, *and* the code of your existing input system is
maintainable, then bidi text input is not          THAT            bad.

There are still some small issues, but I'm going to call this victory,
it's definitely workable already.

[pre13]