Add fallback for table.clear
It turns out not every build of LÖVE supports LuaJIT 2.1's table.clear.
On Linux, it worked for me even on 0.9.1, but it didn't work on 11.3 on
Windows.

The fallback implementation simply iterates over all keys and nils
them, which is fine because my uses of table.clear aren't really for
performance reasons anyway (it's just because I needed to reset tables
while keeping other references to them).

[pre14]