Fix handling of differing fallback font dimensions
If the fallback font doesn't have the same glyph size as the main font
it's used for (like the 8x8 button font being used for the 8x12
Japanese font), then the 8x8 glyphs should be centered within the space
taken up by a main font glyph. This now works properly.

[pre14]