Determine RTL property in font based on context
"Context" here means whether the specific font was invoked as the UI
font, or the level font, or neither of the two. This involves font_ui
and font_level setting an attribute on the font object, but only for
functions that expect it and reset it again. Which is kinda annoying
(read the comments), but I don't immediately see another way to be able
to do it that doesn't involve font object duplication or even more
metatables. It'll do.

[pre14]