Remember latest set level font for new levels
VVVVVV has similar behavior, but also defaults to the font that belongs
to the current language setting unless changed to a font that's not for
the current language. In Ved, this won't work because Ved doesn't have
the same languages that VVVVVV does. But maybe I could add a separate
"default level font" option to the language screen later, who knows.

[pre14]