Make Windows HTTPS code more optimized
Data is downloaded in chunks of 50KB, but they are concatenated at the
end of a Lua string, which means there will be a blowup of memory usage
for large downloads. This commit switches to the better (but not ideal)
tactic of adding all chunks to a table, and then concatenating that at
the end.

[pre20]