Remove textinput UI callback
This is completely unused within Ved (as well as plugins), and I don't
see a reason to use this callback over just using the input system.

The existence of this callback is a little bit in the way of improving
IME support, since we need a central way to tell the OS whether text
input is supposed to be active or not.

The only exception is something like the coordsdialog, but such a thing
could be implemented with the input system anyway.

[pre22]