Enable love.textinput only when necessary
Text input events have normally *always* been called when pressing keys
on your keyboard, even if there was nothing that was actually expecting
text. This didn't really matter - the typed characters would just be
dropped if nothing was doing anything with them.

However, to support IME's properly, the system has to know when input
is expected. Otherwise, if you navigate to a text field using
alphanumeric keyboard shortcuts for example, it can't know when you
intended to start composing a word in Chinese or Japanese.

So now, we use love.keyboard.setTextInput(enable) to enable text input
specifically whenever anything (the new input system, the old input
system, etc) is expecting input, and to disable it when nothing is.

I also fixed the "TAKING INPUT" indicator in debug mode to show up in
the same conditions, so it's not solely dependent on the old
takinginput.

[pre23]