Implement love.textedited
This _should_ implement support for IMEs for CJK, by showing what the
currently edited text is. This is still a little untested, since I
want to do that on another computer and it's easier to just push
immediately and pull there, than to copy it over with a USB stick or
something, heh.

This also adds the behavior change that while uncommitted IME text is
currently set, regular key input (love.keypressed and
love.keyboard.isDown()) is ignored. This is because inputs that should
go to the IME (like Enter, Esc, etc) were also interpreted by Ved,
meaning if you press Enter to commit text, Ved would first insert a
new line and then commit the text. At least, this would happen in
LÖVE 11.4 and below, because in 11.5 the behavior changed to one I
actually need, lol: https://github.com/love2d/love/issues/2031

[pre23]