Tweak IME textedited display
The cursor is now white instead of purple, the text now doesn't blink,
is now positioned correctly, and yeah that was pretty much it.

In theory, if the "selection length" is more than 0, we show an orange
selection background, but that doesn't work on Windows, on which I'm
testing this.

[pre24]