Start 1.11.2-pre01
I'm also going to make a dev-2.0 branch soon - so we can add some
pretty major things like the new theme system (and the new theme), the
new XML parser, and more that's planned! The master branch will be kept
for 1.11.2 and maybe later versions, until 2.0 features are stable
enough that they're no longer really in the way. (At that point I could
always still make a temporary branch for backports in case anything
urgent still comes up.)

Though, right now it's looking pretty good to be able to focus on 2.0:
VVVVVV 2.4 support has just been nearly completed in 1.11.1, and for
the biggest of the two things still left (editing custom special
roomnames) I wanted to wait until the new XML parser was implemented.
I also have quite some free time right now, and Ally just made a new
theme system which fits right in with the plan to finally implement the
shinier theme (while keeping the old theme as an option).

By the way, yeah, 2.0 seems like a pretty colossal jump! But it's not
going to be a rewrite or anything - I'm just taking the chance to
bundle a couple of big features and change the version number from a
3-number 1.x.y system to a 2-number x.y one, continuing at 2.0. The big
new features in 2.0 _are_ probably going to make it the biggest Ved
update ever, but as far as the old version numbering system is
concerned, after getting all the way from 1.0.0 to 1.11.x, we were
gonna stay on 1.xx.x forever if there wasn't going to be a complete
rewrite (which I don't think is likely, I'd rather just keep improving
the existing version). And since 1.10.0, Ved's version numbers have
gotten a little bit harder to read and compare. So I think it's time to
simplify them.

So the plan is: (old plan -> new plan):

1.11.0 -> 1.11.0
1.12.0 -> 2.0
1.13.0 -> 3.0
1.13.1 -> 3.1