Split fused_other into fused_lin and fused_unknown
This is about the distribution method for the update check, which will
mainly be important for auto-updating later, but there have also been
some cases where only some platforms got a minor "b" patch that the
update checker should report for only those platforms (like 1.10.1).

I made the distinction between fused_win, fused_mac and fused_other
back when I thought the Linux version couldn't be fused (in a cross-
distro way). But now the AppImage has existed for a while, and since
I forgot to split that, AppImage versions have been reporting as
fused_other to the update checker. So "other" is practically "linux"
(even though it could technically be legitimately unknown), and new
versions will report as Linux, or "fused_unknown" if unknown.

[pre01]