Add start of autoupdater code for Windows
This adds the source code of the program that Ved will use on Windows
to update its files, and the minimum changes required in Ved itself to
test that separate program (assuming all the update files are already
in place, because Ved doesn't download them yet)

The general idea will be: Ved downloads an update package, verifies its
hash, then unpacks it and starts the separate update program (which is
in the update package). Once Ved itself has quit, the update program
can start replacing files. If successful, the updater program boots Ved
again, and Ved will delete the updater program.

If the updater program runs into an error while replacing files (which
hopefully never happens!) then you get the option to retry the failed
operation, or roll back the entire update, which makes it revert all
changes in reverse order.

Of course, macOS and Linux support will follow as well, as well as
support for downloading a new .love. Updating via git pull would be
nice too, but let's see how this goes first.

[pre02]