Don't place # in empty setroomname/setactivitytext
If you want to set the roomname to be empty, you'd use the setroomname
command with a blank line below it. But Ved inserts a # on that line,
because it doesn't know that doesn't have to in that case. Thanks to
Reese for reporting and verifying the fix!

The same probably applies to setactivitytext, so I've added that too.

[pre04]