Improve alignments in dialogs for RTL
Dialogs which only have text (and no form fields) are now right-aligned
in RTL languages.

I also removed some trailing spaces from strings that ended with
colons, so that the space wouldn't push the label further right when
left-aligned in a dialog box.

So:

|:name
|[         ]

is better than:

| :name
|[         ]

(In the music editor I actually added one because strings were both
being used for concatenation and as field labels, but other than that
I removed the spaces from the field labels)

[pre09]