Put remaining input labels right way around in RTL
The level load and search screens have text input that's preceded by
a label like "Load this level:". In Arabic, that should logically be
flipped around (so the text input is to the left of the label), but
most ways of doing that would be problematic in ways I'm currently too
lazy to explain in detail since the interface isn't fully flipped
around (things like, how much space do you give the input text, or do
you want to break alignment of input names with the names in the table
above).

The solution that I think looks pretty nice is to make the label move
along with the input text, but darken it enough that it doesn't look
out of place. It seemed to look kinda weird if the label stays white
and just moves along, so the solution was going to be to just remove
the label if there was any text typed in, but making it darkened is
close to that visually while not having the disadvantage of removing
the label completely.

[pre09]