Improve font context (ui vs level) recognition
This was going to be a convoluted solution that I puzzled together with
all my lua experience, and then I realized it could be much simpler...

So. There's the cVedFont class which represents a font itself. There's
a font object for several writing systems in the fonts_main table, and
there's fonts_custom for font objects belonging to the loaded level.

Most of the time, to access a specific font, you'd be using font_ui,
font_level or font_8x8, which _behave_ like a font object, but actually
proxy everything to one of the *actual* font objects, depending on what
the UI font or level font is currently set to (you could have an
English level in Arabic Ved, or an Arabic level in English Ved, and the
same fonts are used in both cases but the roles are reversed).

The problem is, the fonts themselves sometimes need to know what role
they're currently being asked to play - i.e. whether you're calling
font_ui:print() or font_level:print(). This matters because the UI font
and level font should sometimes act a little differently - particularly
whether to look at Ved's language RTL setting or the level RTL setting.

I had solved this by setting a context attribute on the font object
when using font_ui/font_level - the proxy would tell the font what role
it has to play, and once the font was done the context would then need
to be cleaned up again, so if the font is later used directly (not
through a font_ui/font_level/etc proxy), it doesn't still think it's
being proxied.

But the proxy can't unset the context attribute! In __index, it can
prepare for a method on the real font object to be called, but after
it returns the requested function, it can't do anything and doesn't
have control when it's actually called. So, the font has to clean that
up by itself, every time. I didn't want to do that on every cVedFont
method, so I made a whitelist of methods that the context attribute is
prepared for because they need it and also know how to clean it up.

This is ugly and has tons of problems, and I don't really know how my
font_ui:is_rtl() calls were seemingly working fine because is_rtl was
not a whitelisted function to get the context even though it needed it,
because I thought is_rtl would only be called from one of the
whitelisted functions, but it still worked when called directly. Oh
well, there's probably a reason for that, to do with how the metatables
are set up that I can realize now if I think about it for a bit. I
couldn't whitelist is_rtl because it was being called from other
whitelisted methods, so it would clean up too early when the caller
still needed it.

I eventually came up with a solution where the __index metamethod
wouldn't return the requested method - say, font["font_ja"].print -
but would return a special wrapper function, like font.context_call.
That wrapper calls the *actual* function, but it has control both
before and after, so it could set and clean up the context attribute.

This... doesn't require a specific whitelist, but causes problems if
you chain multiple function calls in the same expression - think
font_ui:print(...font_ui:getWidth()...) - because __index would first
be called twice (thus prepare context_call twice) and then context_call
would just call the same function twice, instead of two different ones.

So, a further extension to that would be to make a unique wrapper
function for every possible method for every context. So, a function
for `print` being called in level context, for example. That wrapper
function would then set and unset the context. This could be done
dynamically - __index knows the name of any function that's being
called, so it could keep a table with all the functions that have been
called in all the contexts, and make all of them distinct and
unchanging.

Then I had a realization that a metatable's __index would only be
called if table[index] didn't exist in the main table (font_ui or
font_level), and **the self passed to cVedFont method is actually
font_ui or font_level if you call them through that**...

Of course, so, if you call font_ui:print(...), and font_ui proxies it
to fonts_main["font_ja"]:print(...), you're not actually calling:

fonts_main["fonts_ja"].print(fonts_main["fonts_ja"], ...)

But you're calling this instead:

fonts_main["fonts_ja"].print(font_ui, ...)

Thus, everything that the print function wants to get from its own
object/class (self), that seems completely internal, is actually _also_
being proxied, and it's being calculated in __index which font to use!
(Still the same one, hopefully)

So, if you simply add a context attribute to all of the proxy fonts,
the internal cVedFont methods can just see it. And nothing has to be
prepared or cleaned up.

[pre09]